+++ Sources/Swift/Integrations/Performance/IO/FileManager+SentryTracing.swift
@@ -12,7 +12,7 @@ public extension FileManager {
     ///
     /// - Important: Using this method with auto-instrumentation for file operations enabled can lead to duplicate spans on older operating system versions.
     ///              It is recommended to use either automatic or manual instrumentation. You can disable automatic instrumentation by setting
-    ///              `options.enableSwizzling` to `false` when initializing Sentry.
+    ///              `options.experimental.enableFileManagerSwizzling` to `false` when initializing Sentry.
     /// - Parameters:
     ///   - path: The path for the new file.
     ///   - data: A data object containing the contents of the new file.
@@ -28,7 +28,7 @@ public extension FileManager {
                 atPath: path,
                 contents: data,
                 attributes: attr,
-                origin: SentryTraceOrigin.manualFileData) { path, data, attr in
+                origin: SentryTraceOriginManualFileData) { path, data, attr in
                     self.createFile(atPath: path, contents: data, attributes: attr)
             }
     }
@@ -37,13 +37,13 @@ public extension FileManager {
     ///
     /// - Important: Using this method with auto-instrumentation for file operations enabled can lead to duplicate spans on older operating system versions.
     ///              It is recommended to use either automatic or manual instrumentation. You can disable automatic instrumentation by setting
-    ///              `options.enableSwizzling` to `false` when initializing Sentry.
-    /// - Parameter url: A file URL specifying the file or directory to remove. 
+    ///              `options.experimental.enableFileManagerSwizzling` to `false` when initializing Sentry.
+    /// - Parameter url: A file URL specifying the file or directory to remove.
     ///                  If the URL specifies a directory, the contents of that directory are recursively removed.
     /// - Note: See ``FileManager.removeItem(at:)`` for more information.
     func removeItemWithSentryTracing(at url: URL) throws {
         let tracker = SentryFileIOTracker.sharedInstance()
-        try tracker.measureRemovingItem(at: url, origin: SentryTraceOrigin.manualFileData) { url in
+        try tracker.measureRemovingItem(at: url, origin: SentryTraceOriginManualFileData) { url in
             try self.removeItem(at: url)
         }
     }
@@ -52,13 +52,13 @@ public extension FileManager {
     ///
     /// - Important: Using this method with auto-instrumentation for file operations enabled can lead to duplicate spans on older operating system versions.
     ///              It is recommended to use either automatic or manual instrumentation. You can disable automatic instrumentation by setting
-    ///              `options.enableSwizzling` to `false` when initializing Sentry.
-    /// - Parameter path: A path string indicating the file or directory to remove. 
+    ///              `options.experimental.enableFileManagerSwizzling` to `false` when initializing Sentry.
+    /// - Parameter path: A path string indicating the file or directory to remove.
     ///                   If the path specifies a directory, the contents of that directory are recursively removed.
     /// - Note: See ``FileManager.removeItem(atPath:)`` for more information.
     func removeItemWithSentryTracing(atPath path: String) throws {
         let tracker = SentryFileIOTracker.sharedInstance()
-        try tracker.measureRemovingItem(atPath: path, origin: SentryTraceOrigin.manualFileData) { path in
+        try tracker.measureRemovingItem(atPath: path, origin: SentryTraceOriginManualFileData) { path in
             try self.removeItem(atPath: path)
         }
     }
@@ -69,7 +69,7 @@ public extension FileManager {
     ///
     /// - Important: Using this method with auto-instrumentation for file operations enabled can lead to duplicate spans on older operating system versions.
     ///              It is recommended to use either automatic or manual instrumentation. You can disable automatic instrumentation by setting
-    ///              `options.enableSwizzling` to `false` when initializing Sentry.
+    ///              `options.experimental.enableFileManagerSwizzling` to `false` when initializing Sentry.
     /// - Parameters:
     ///   - srcURL: The file URL that identifies the file you want to copy.
     ///             The URL in this parameter must not be a file reference URL.
@@ -78,7 +78,7 @@ public extension FileManager {
     /// - Note: See ``FileManager.copyItem(at:to:)`` for more information.
     func copyItemWithSentryTracing(at srcURL: URL, to dstURL: URL) throws {
         let tracker = SentryFileIOTracker.sharedInstance()
-        try tracker.measureCopyingItem(at: srcURL, to: dstURL, origin: SentryTraceOrigin.manualFileData) { srcURL, dstURL in
+        try tracker.measureCopyingItem(at: srcURL, to: dstURL, origin: SentryTraceOriginManualFileData) { srcURL, dstURL in
             try self.copyItem(at: srcURL, to: dstURL)
         }
     }
@@ -87,7 +87,7 @@ public extension FileManager {
     ///
     /// - Important: Using this method with auto-instrumentation for file operations enabled can lead to duplicate spans on older operating system versions.
     ///              It is recommended to use either automatic or manual instrumentation. You can disable automatic instrumentation by setting
-    ///              `options.enableSwizzling` to `false` when initializing Sentry.
+    ///              `options.experimental.enableFileManagerSwizzling` to `false` when initializing Sentry.
     /// - Parameters:
     ///   - srcPath: The path to the file or directory you want to move. 
     ///   - dstPath: The path at which to place the copy of `srcPath`. 
@@ -95,7 +95,7 @@ public extension FileManager {
     /// - Note: See ``FileManager.copyItem(atPath:toPath:)`` for more information.
     func copyItemWithSentryTracing(atPath srcPath: String, toPath dstPath: String) throws {
         let tracker = SentryFileIOTracker.sharedInstance()
-        try tracker.measureCopyingItem(atPath: srcPath, toPath: dstPath, origin: SentryTraceOrigin.manualFileData) { srcPath, dstPath in
+        try tracker.measureCopyingItem(atPath: srcPath, toPath: dstPath, origin: SentryTraceOriginManualFileData) { srcPath, dstPath in
             try self.copyItem(atPath: srcPath, toPath: dstPath)
         }
     }
@@ -104,7 +104,7 @@ public extension FileManager {
     ///
     /// - Important: Using this method with auto-instrumentation for file operations enabled can lead to duplicate spans on older operating system versions.
     ///              It is recommended to use either automatic or manual instrumentation. You can disable automatic instrumentation by setting
-    ///              `options.enableSwizzling` to `false` when initializing Sentry.
+    ///              `options.experimental.enableFileManagerSwizzling` to `false` when initializing Sentry.
     /// - Parameters:
     ///   - srcURL: The file URL that identifies the file or directory you want to move. 
     ///             The URL in this parameter must not be a file reference URL.
@@ -116,7 +116,7 @@ public extension FileManager {
         try tracker.measureMovingItem(
             at: srcURL,
             to: dstURL,
-            origin: SentryTraceOrigin.manualFileData) { srcURL, dstURL in
+            origin: SentryTraceOriginManualFileData) { srcURL, dstURL in
             try self.moveItem(at: srcURL, to: dstURL)
         }
     }
@@ -125,7 +125,7 @@ public extension FileManager {
     ///
     /// - Important: Using this method with auto-instrumentation for file operations enabled can lead to duplicate spans on older operating system versions.
     ///              It is recommended to use either automatic or manual instrumentation. You can disable automatic instrumentation by setting
-    ///              `options.enableSwizzling` to `false` when initializing Sentry.
+    ///              `options.experimental.enableFileManagerSwizzling` to `false` when initializing Sentry.
     /// - Parameters:
     ///   - srcPath: The path to the file or directory you want to move.
     ///   - dstPath: The new path for the item in `srcPath`. 
@@ -133,7 +133,7 @@ public extension FileManager {
     /// - Note: See ``FileManager.moveItem(atPath:toPath:)`` for more information.
     func moveItemWithSentryTracing(atPath srcPath: String, toPath dstPath: String) throws {
         let tracker = SentryFileIOTracker.sharedInstance()
-        try tracker.measureMovingItem(atPath: srcPath, toPath: dstPath, origin: SentryTraceOrigin.manualFileData) { srcPath, dstPath in
+        try tracker.measureMovingItem(atPath: srcPath, toPath: dstPath, origin: SentryTraceOriginManualFileData) { srcPath, dstPath in
             try self.moveItem(atPath: srcPath, toPath: dstPath)
         }
     }