+++ Sources/Swift/Integrations/Performance/IO/SentryFileIOTracker+SwiftHelpers.swift
@@ -40,7 +40,7 @@ extension SentryFileIOTracker {
         guard url.scheme == NSURLFileScheme else {
             return try method(data, url, options)
         }
-        guard let span = self.span(forPath: url.path, origin: origin, operation: SentrySpanOperation.fileWrite, size: UInt(data.count)) else {
+        guard let span = self.span(forPath: url.path, origin: origin, operation: SentrySpanOperationFileWrite, size: UInt(data.count)) else {
             return try method(data, url, options)
         }
         defer {
@@ -60,7 +60,7 @@ extension SentryFileIOTracker {
         guard url.scheme == NSURLFileScheme else {
             return try method(url)
         }
-        guard let span = self.span(forPath: url.path, origin: origin, operation: SentrySpanOperation.fileDelete) else {
+        guard let span = self.span(forPath: url.path, origin: origin, operation: SentrySpanOperationFileDelete) else {
             return try method(url)
         }
         defer {
@@ -74,7 +74,7 @@ extension SentryFileIOTracker {
         origin: String,
         method: (_ path: String) throws -> Void
     ) rethrows {
-        guard let span = self.span(forPath: path, origin: origin, operation: SentrySpanOperation.fileDelete) else {
+        guard let span = self.span(forPath: path, origin: origin, operation: SentrySpanOperationFileDelete) else {
             return try method(path)
         }
         defer {
@@ -91,12 +91,12 @@ extension SentryFileIOTracker {
         method: (_ path: String, _ data: Data?, _ attributes: [FileAttributeKey: Any]?) -> Bool
     ) -> Bool {
         let size = UInt(data?.count ?? 0)
-        guard let span = self.span(forPath: path, origin: origin, operation: SentrySpanOperation.fileWrite, size: size) else {
+        guard let span = self.span(forPath: path, origin: origin, operation: SentrySpanOperationFileWrite, size: size) else {
             return method(path, data, attr)
         }
         defer {
             if let data = data {
-                span.setData(value: data.count, key: SentrySpanDataKey.fileSize)
+                span.setData(value: data.count, key: SentrySpanDataKeyFileSize)
             }
             span.finish()
         }
@@ -115,7 +115,7 @@ extension SentryFileIOTracker {
         guard srcUrl.scheme == NSURLFileScheme && dstUrl.scheme == NSURLFileScheme else {
             return try method(srcUrl, dstUrl)
         }
-        guard let span = self.span(forPath: srcUrl.path, origin: origin, operation: SentrySpanOperation.fileCopy) else {
+        guard let span = self.span(forPath: srcUrl.path, origin: origin, operation: SentrySpanOperationFileCopy) else {
             return try method(srcUrl, dstUrl)
         }
         defer {
@@ -130,7 +130,7 @@ extension SentryFileIOTracker {
         origin: String,
         method: (_ srcPath: String, _ dstPath: String) throws -> Void
     ) rethrows {
-        guard let span = self.span(forPath: srcPath, origin: origin, operation: SentrySpanOperation.fileCopy) else {
+        guard let span = self.span(forPath: srcPath, origin: origin, operation: SentrySpanOperationFileCopy) else {
             return try method(srcPath, dstPath)
         }
         defer {
@@ -151,7 +151,7 @@ extension SentryFileIOTracker {
         guard srcUrl.scheme == NSURLFileScheme && dstUrl.scheme == NSURLFileScheme else {
             return try method(srcUrl, dstUrl)
         }
-        guard let span = self.span(forPath: srcUrl.path, origin: origin, operation: SentrySpanOperation.fileRename) else {
+        guard let span = self.span(forPath: srcUrl.path, origin: origin, operation: SentrySpanOperationFileRename) else {
             return try method(srcUrl, dstUrl)
         }
         defer {
@@ -166,18 +166,18 @@ extension SentryFileIOTracker {
         origin: String,
         method: (_ srcPath: String, _ dstPath: String) throws -> Void
     ) rethrows {
-        guard let span = self.span(forPath: srcPath, origin: origin, operation: SentrySpanOperation.fileRename) else {
+        guard let span = self.span(forPath: srcPath, origin: origin, operation: SentrySpanOperationFileRename) else {
             return try method(srcPath, dstPath)
         }
         defer {
             span.finish()
         }
         try method(srcPath, dstPath)
-        guard let span = self.span(forPath: url.path, origin: origin, operation: SentrySpanOperationFileWrite, size: UInt(data.count)) else {
-            return try method(data, url, options)
+        guard let span = self.span(forPath: srcPath, origin: origin, operation: SentrySpanOperationFileRename) else {
+            return try method(srcPath, dstPath)
         }
         do {
-            try method(data, url, options)
+            try method(srcPath, dstPath)
             span.finish()
         } catch {
             span.finish(status: .internalError)