+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Features
 
-- Add extension for `FileManager` to track file I/O operations with Sentry (#4863)
 - Add extension for `Data` to track file I/O operations with Sentry (#4862)
+- Add extension for `FileManager` to track file I/O operations with Sentry (#4863)
 
 ## 8.46.0